Implement TCB model as a plugin
