+++ src/plugins/index.js
@@ -1,1 +1,10 @@
 // This file will contains exported plugin collections
+
+import TapChangerBlockingModelPlugin from './automaton/model/TapChangerBlockingModelPlugin';
+
+export const AutomatonModelGroupPlugins = [
+    {
+        id: 'TapChangerBlockingModelPlugin',
+        modePlugin: TapChangerBlockingModelPlugin,
+    },
+];
