+++ src/utils/automata.js
@@ -5,7 +5,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-import { AutomatonProperties } from '../constants/equipmentDefinition';
+import { AutomatonProperties } from '../constants/automatonDefinition';
 
 export const getAutomatonPropertiesByModel = (model) => {
     return AutomatonProperties[model];
